Start adding to pr.3
MUCH work has to be done and that will likely not be done for a while,
probably not until after IOCCC29, but to get my directory clean I am
committing this. Only one function has been described but the
DESCRIPTION has the macros and functions listed now (that were not
already listed). The HISTORY section was elaborated slightly as well.

Also updated LICENSE to match pr.c.